273: Added SidecarConfig to component and service config
