+++ src/dependency-manager/src/config/sidecar-config.ts
@@ -1,0 +1,8 @@
+import { LivenessProbeConfig } from './common-config';
+import { ResourceConfig } from './resource-config';
+
+export interface SidecarConfig extends ResourceConfig {
+  enabled: boolean;
+  liveness_probe?: LivenessProbeConfig;
+  debug?: ResourceConfig;
+}
